requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-OAuth==0.12
MarkupSafe==0.18
boto==2.35.1
PyYAML==3.11
Werkzeug==0.9.4
clay-flask==2.1.5
httplib2==0.9
itsdangerous==0.23
oauth2==1.5.211
oursql
pytz==2013.9
webassets==0.8
wsgiref==0.1.2
python-dateutil
psycopg2
